backend/requirements.txt
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.4.0
billiard==3.6.4.0
celery==5.2.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.3
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-nested-routers==0.93.4
idna==3.3
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.1
kombu==5.2.4
MarkupSafe==2.1.1
model-bakery==1.5.0
mysqlclient==2.1.0
oauthlib==3.2.0
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
prompt-toolkit==3.0.30
py==1.11.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
python3-openid==3.2.0
pytz==2022.1
redis==4.3.4
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
pytest-mock==3.8.2